requirements.txt
asgiref==3.6.0
attrs==22.2.0
Brotli==1.0.9
cffi==1.15.1
cssselect2==0.7.0
Django==4.1.3
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
fonttools==4.39.3
html5lib==1.1
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
psycopg2==2.9.5
pycparser==2.21
pydyf==0.6.0
PyJWT==2.6.0
pyphen==0.14.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-environ==0.4.54
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
typing_extensions==4.5.0
uritemplate==4.1.1
weasyprint==58.1
webencodings==0.5.1
zopfli==0.2.2
